requirements.txt
numpy==1.15.0
matplotlib==2.2.2
scipy==1.1.0
rasterio==1.0.20
pandas==0.23.4
geopandas==0.4.0
pydeck==0.1.dev5
seaborn==0.9.0
PyYAML==5.1.2
GDAL==2.4.1
